+++ pajbot/apiwrappers/seventv.py
@@ -20,6 +20,8 @@ class SevenTVAPI(BaseAPI):
                     code=emote_data["name"],
                     provider="7tv",
                     id=emote_id,
+                    max_width=emote_data["width"][3],
+                    max_height=emote_data["height"][3],
                     urls={
                         "1": get_url(emote_id, "1"),
                         "2": get_url(emote_id, "2"),
@@ -37,6 +39,8 @@ query fetchGlobalEmotes($query: String!, $globalState: String, $page: Int, $limi
     search_emotes(query: $query, globalState: $globalState, page: $page, limit: $limit, pageSize: $pageSize) {
         id
         name
+        width
+        height
     }
 }"""
 
@@ -82,6 +86,8 @@ query fetchUserEmotes($id: String!) {
         emotes {
             id
             name
+            width
+            height
         }
     }
 }"""