+++ pajbot/constants.py
@@ -1,1 +1,1 @@
-VERSION = "1.60"
+VERSION = "1.61"