+++ pajbot/dispatch.py
@@ -17,7 +17,7 @@ class Dispatch:
 
     @staticmethod
     def add_win(bot, source, message, event, args):
-        # XXX: this is ugly as fuck
+        log.warning("DEPRECATED - Use the $(increasekvi:br_wins) variable")
         bot.kvi["br_wins"].inc()
         bot.me(f"{source} added a BR win!")
 
@@ -34,42 +34,38 @@ class Dispatch:
         # Make sure we got both an alias and a response
         message_parts = message.split()
         if len(message_parts) < 2:
-            bot.whisper(source, "Usage: !add command ALIAS [options] RESPONSE")
+            bot.send_message_to_user(source, "Usage: !add command ALIAS [options] RESPONSE", event, method="whisper")
             return False
 
         options, response = bot.commands.parse_command_arguments(message_parts[1:])
 
         if options is False:
-            bot.whisper(source, "Invalid command")
+            bot.send_message_to_user(source, "Invalid command", event, method="whisper")
             return False
 
         options["added_by"] = source.id
 
         alias_str = message_parts[0].replace("!", "").lower()
-        type = "say"
-        if options["whisper"] is True:
-            type = "whisper"
-        elif options["reply"] is True:
-            type = "reply"
-        elif response.startswith("/me") or response.startswith(".me"):
-            type = "me"
+        action_type: str = options.get("action_type", "say")
+        if response.startswith("/me") or response.startswith(".me"):
+            action_type = "me"
             response = " ".join(response.split(" ")[1:])
-        elif options["whisper"] is False or options["reply"] is False:
-            type = "say"
-        action = {"type": type, "message": response}
+        action = {"type": action_type, "message": response}
 
         command, new_command, alias_matched = bot.commands.create_command(alias_str, action=action, **options)
         if new_command is True:
-            bot.whisper(source, f"Added your command (ID: {command.id})")
+            bot.send_message_to_user(source, f"Added your command (ID: {command.id})", event, method="whisper")
 
             log_msg = f"The !{command.command.split('|')[0]} command has been created"
             AdminLogManager.add_entry("Command created", source, log_msg)
             return True
 
         # At least one alias is already in use, notify the user to use !edit command instead
-        bot.whisper(
+        bot.send_message_to_user(
             source,
             f"The alias {alias_matched} is already in use. To edit that command, use !edit command instead of !add command.",
+            event,
+            method="whisper",
         )
         return False
 
@@ -210,7 +206,7 @@ class Dispatch:
 
     @staticmethod
     def remove_win(bot, source, message, event, args):
-        # XXX: This is also ugly as fuck
+        log.warning("DEPRECATED - Use the $(decreasekvi:br_wins) variable")
         bot.kvi["br_wins"].dec()
         bot.me(f"{source} removed a BR win!")