+++ pajbot/web/routes/clr.py
@@ -15,17 +15,3 @@ log = logging.getLogger("pajbot")
 @nocache
 def overlay(widget_id, **options):
     return render_template("clr/overlay.html", widget={})
-
-
-@page.route("/fatoverlay/<widget_id>")
-@page.route("/fatoverlay/<widget_id>/<random_shit>")
-@nocache
-def fatoverlay(widget_id, **options):
-    return render_template("clr/fatoverlay.html", widget={})
-
-
-@page.route("/crazyoverlay/<widget_id>")
-@page.route("/crazyoverlay/<widget_id>/<random_shit>")
-@nocache
-def crazyoverlay(widget_id, **options):
-    return render_template("clr/crazyoverlay.html", widget={})